# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: CoverPolikrami API
  version: 0.1.0
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, confirmPassword, acceptTerms, acceptPrivacy]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                confirmPassword: { type: string }
                name: { type: string }
                role: { type: string, enum: [user, designer] }
                acceptTerms: { type: boolean }
                acceptPrivacy: { type: boolean }
                acceptRevenueShare: { type: boolean }
      responses:
        '201': { description: Created }
...